Revert line height tweak to old patching method
EndShadowBar calls NewLine(22).  Since that specific NewLine call is not directly intercepted, changing the minHeight in the intercepted DrawAction can't affect it.

Revert back to patching NewLine, which will always adjust the minHeight.  This has the downside of doing extra work, since it will execute on both intercept and  draw calls.