Pool InfoCards
The bulk of the memory savings from this don't actually come from saving on the InfoCard constructor, but from preserving the collection capacities.

If the current frame is the same as the last frame, the collections are guaranteed to be correctly sized already, and don't need to be resized at all.