Constrain text length and flex UI size around it
Edit the LocText to truncate and left align.  Add a LayoutElement to it so that its width can be directly set.  Locking the layout prevents the rest of the UI system from overriding what is set.

This is superior to grid constraining, because grid constraining doesn't allow LocText to truncate.  Strings could be manually truncated, but then the truncation length would have to be set separate from the grid width, when they're actually logically dependent values.