Add dragging to mod entries
Create a custom ADragMe component, similar to the base-game's DragMe.  Reimplementation is necessary because the original is not flexible enough to handle this use case, and is not easily extendable.

Broke the logic to constrain text length on mod entries out into a dedicated method, rather than an anonymous function.

DragListener currently does nothing.