Change color of the entry while it is being dragged
When dragging a mod entry, the original entry is left so that users can identify where it is placed.  This changes the color of that panel to distinguish it from entries that are not being dragged.

Minor UI appearance tweaks to make testing more tolerable: more space between entries, and better entry colors.