Fix incorrectly disabled entries when jumping to the top in virtual scroll
When jumping to the top (clicking on the scroll bar), the lower entries would not be properly displayed due to the index checks.

Checking first and last indices against the prior update would have worked.  However, since the time saved with the early out is very minor, it was removed to simplify the code.