Use Unity's messaging system to update children on search changes
Unity's messaging system is more flexible than passing around references through the entire UI hierarchy.

Define a wrapper method that allows an event to be easily executed on an entire GO hierarchy.

Event target interfaces and execution logic are defined in the triggering class.  Concrete targets are defined on the listening class, and implement the modification logic through an instance.