Add extracted GOs to the info panel
Extract any GOs on the base-game mod entry UI that are not present on its prefab.  These are re-parented to a panel in the info panel, allowing mod manager to flexibly handle any mod (theoretically).

Change to re-using the same GO instead of re-instantiating.  This preserves the full state of the GO, so that all of its capabilities carry into the new UI.