Fix: Labels not applied to all buildings
Switch from patching delegates and using SaveLoadRoot to patching
methods and using ILogicUIElement.

Necessary since a game update moved some buildings out of being iterated
with SaveLoadRoot. That iteration only includes buildings that are
highlighted in the overlay, but labels need to be applied to
non-highlighted buildings (ex: smart battery).

Switching to ILogicUIElement allows access to everything that draws
logic ports to the overlay. This approach simplifies the patches, but
requires more complicated logic to find the GO.