Fix: Buildings over logic gates do not have labels
Add labels to buildings with an automation cell overlapping a logic
gate.

Multiple buildings with logic ports can overlap each other on different
layers.  Using the first building found at that cell can yield a
building that does not correspond to the cell.  To always choose the
right building, it is necessary to check that it contains the cell.