Fix: Labels applied to incorrect buildings
Allow filtering label application based on the type of logic ports that
the target object has.  This removes unnecessary labels from the
massage table, automatic dispenser, and spice grinder.

Some objects have settings and logic ports, but it is not internally
consistent to display the labels.  The intent is to apply labels to any
settings that change what a port would output.  Labels should only be
applied to input ports when the setting alters how they might be wired.
This is why the filter settings are applied to inputs and outputs.
Some objects will still be exceptions, necessitating a blocklist.