Fix: Fast clear crash when closing categories
Update 544519 introduced build menu changes that cause crashes in No
Research Alerts.

Update reference to ActiveCategoryBuildingToggles.  The property was
changed to a field, causing a missing property crash.

Access the category through the toggle info.
ActiveCategoryBuildingToggles is no longer limited to toggles in the
currently open category - it includes all buildings (it is not clear if
this is an unintentional error).

Apply patch to CloseCategoryPanel.  When ClearButtons is called, the
toggle info is null.  CloseCategoryPanel also more closely represents
the intent of the patch.