Fix: Suppress button description missing items
Add notifications to the Suppress Current button's description
immediately after they are cleared.

When notifications are unsuppressed, RefreshNotifications() is called
to remove and re-add the notifications.  Removal is instantly tracked
through the Notifier.Remove patch, but adding puts notifications into a
pending state that isn't tracked at NotificationScreen.AddNotification
until the next Update.  When the button tooltip is updated, the
notification is not tracked and is omitted from the description.

Switching from NotificationScreen.AddNotification to Notifier.Add
enables these to be immediately tracked.  Because the GO cannot be
accessed before the method runs, the NotificationInterceptor has to be
broken out into a separate patch.