Refac: Support singletons for string registration
Change fieldless string registration to use singletons.  Change
standard string registration to easily use singletons if needed in the
future.

Future features will require the ability to get additional information
from fieldless string types.  The easiest way to achieve this is
through abstract/interface methods, which can not be static and thus
require an instance reference.

Due to limitations of the runtime, static/default interface members are
not supported, greatly limiting the approaches that can work.