Feat: Custom POT Generator
Re-implement Klei's POT generator.  Remove the useless comments left on
each line and fix Klei's dictionary concat to actually work.

This could have been done through patches or reverse patches, but
enough modifications are required that it is simpler to fully redefine.
This will only be used for debug builds and is standalone code, so this
isn't a terrible approach.

This is largely copied from Klei's decompiled code and has been renamed
and broken into methods for clarity.