Feat: Generate POT for all fieldless strings
Include fieldless strings that do not have a default implementation in
generated POTs.

This makes it easier for translators to see what is available to
translate and prevents them from needing to manually add entries.

Default implementations can't be public for fieldless strings.  If they
are, they will be included in the default strings tree and the extra
strings tree, leading to duplicate keys when they are concated.  This
is fine since the LocKeys shouldn't be directly referenced for
fieldless strings.