Fix: Solution build error zipping translations
Configure the translations zip task to run for only one project
(AzeLib).

Each project tries to create Translations.zip simultaneously, leading
to issues when building the entire solution.

Also change assembly publicization to execute only for AzeLib.  This
isn't strictly necessary, but should result in faster builds and is
conceptually cleaner.

The downside of this change is that it becomes possible to hit edge
cases where a release build does not rebuild AzeLib and thus doesn't
zip translations.  This can be worked around with a rebuild, but
necessitates checking before committing.

Using After/Before.oniMods.sln.targets to implement these targets would
ensure they are always run once, however it poses other issues:
- Because the target would be outside the "project scope", the package
  reference for the zip task doesn't work.
- Running the after/before targets requires executing MSBuild from the
  command line, or installing an extension for VS.
For now, constraining to the lib is the simplest solution.