Feat: Advanced logging for BIC crashes
Add logging to dump as much information as possible when Better Info
Cards encounters the missing key (#34) or stacked germs (#33) crash.

Breakpoints have to be passed as a parameter because the try block
moves the local var out of scope of the local method.

Try block formatting intentionally incorrect to simplify diffs.