Fix: Screens not closing in certain configs
Close the build menu and overlay screen on tool change if it is set to
close before or at the same time as the selected object is set to
deselect.

The prior implementation would not close, even if the build menu and
selected object had the same setting, if that setting was over 1 (e.g.,
Overlay screen set to 1, selected object and build menu set to 2).