Fix: KeyNotFound crash from improper grouping
Fix: KeyNotFound crash from improper grouping

Fix KeyNotFound crash when trying to get a TextInfo (TI) that does not
exist for a given InfoCard (IC).  This is caused by improperly grouping
ICs with different TIs together, which violates core assumptions made
in the code.  Improper grouping was caused because history was
implicitly shared across each group, instead of only passing down to
recursive calls.

This is fixed by creating new lists that are not the same reference.
Fix #34.

Reproduction requires at least 4 InfoCards.  The first 2 must have
matching titles and TI counts so that they form a partial group that
must be split by TIs.  The second group must also form a partial group,
and must contain at least one differentiating TI (a TI that is not
common between the ICs in the group).  This TI must have been present on
the 1st IC in the 1st group, and it must be a TI with splits defined by
a Converter.

Concrete example (simplified):

GasPipe
Contents: Empty
Temp

GasPipe
Construction Errand
Top Priority

LiquidPipe
Contents: Empty
Temp

LiquidPipe
Construction Errand
Top Priority

The 2 GasPipes have the same title and TI count, so they are split by
TIs.  This saves "Contents: Empty" and "Temp" to history.  Then, BIC
attempts to split the LiquidPipes by TIs.  Because both TIs for the 1st
LiquidPipe are already present in the history, it's assumed that there
are no differentiating TIs left in the group.  This will then crash
once BIC attempts to split the (improper) group into temperature bands
since it expects the given TI to be present on all ICs in the group.