Fix: Deselect functionality
Fix deselect features broken by game updates.

Patch moved to ToolMenu as this receives the key event before the
PlanScreen.

Interface with current selected object through the PlayerController to
handle new cases in the game.

No longer skip the existing OnKeyUp methods.  They have grown and
include multiple other effects.  This is safe as the MouseRight Action
is consumed in this code, so deselect actions should not be duplicated.